requirements.txt
notion-client>=2.0.0
feedparser>=6.0.0
python-dotenv>=1.0.0
anthropic>=0.25.0
